Update to Octicons v6.0.1, regenerate.
The build output format has changed slightly, so we adapt to it. Instead
of build/svg.json with full SVG content, it's now build/data.json with
a JSON object containing the SVG path, width, height, etc. We use that
to generate the entire SVG XML, and parse that as before.

In the future, the generate steps can be combined and simplified, but
that can be done later, when there's more value in doing so.