Rename package to octicon (from octiconssvg).
This is a simpler, better name. Users of the package will see things
like octicon.GitCommit() and octicon.Icon("alert"), which reads better.

The reason why octiconssvg was originally used is because octicons
was taken by the now-deprecated icon-font version of this package,
and I hadn't yet realized that "octicon" would make a better name.